requirements.txt
fastapi
uvicorn
requests
python-dotenv
pytz
google
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
PyPDF2
PyMuPDF
python-docx
python-dotenv
